feat: 일단 api 호출은 가능함 구조 싹 다 갈아버리는 중

+++ Drtaa-Android/core-model/src/main/java/com/drtaa/core_model/network/RequestCarStatus.kt
@@ -1,8 +1,8 @@
 package com.drtaa.core_model.network
 
 data class RequestCarStatus(
-    val rentId: Long,
-    val travelId: Long,
-    val travelDatesId: Long,
-    val datePlacesId: Long,
+    val rentId: Int,
+    val travelId: Int,
+    val travelDatesId: Int,
+    val datePlacesId: Int,
 )
