+++ Drtaa-Android/core-model/src/main/java/com/drtaa/core_model/rent/CarPosition.kt
@@ -3,8 +3,8 @@ package com.drtaa.core_model.rent
 data class CarPosition(
     val latitude: Double,
     val longitude: Double,
-    val rentId: Long,
-    val travelId: Long,
-    val travelDatesId: Long,
-    val datePlacesId: Long
+    val rentId: Int,
+    val travelId: Int,
+    val travelDatesId: Int,
+    val datePlacesId: Int
 )
